requirements.txt
pyqt5
requests
json5
pillow